requirements.txt
flask==3.0.0
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.2
google-cloud-aiplatform==1.68.0
requests==2.32.3
